+++ packages/http/src/mocker/callback/__tests__/callbacks.spec.ts
@@ -10,6 +10,7 @@ describe('runCallback()', () => {
     error: jest.fn(),
     warn: jest.fn(),
     info: jest.fn(),
+    debug: jest.fn(),
   };
 
   beforeEach(() => {
@@ -57,17 +58,8 @@ describe('runCallback()', () => {
         },
       })(logger)();
 
-      expect(fetch).toHaveBeenCalledWith('http://example.com/get/200/5/weird/content', {
-        method: 'get',
-        body: '{"about":"something"}',
-        headers: { 'content-type': 'application/json' },
-      });
-      expect(logger.info).toHaveBeenNthCalledWith(
-        1,
-        { name: 'CALLBACK' },
-        'test callback: Making request to http://example.com/get/200/5/weird/content...'
-      );
-      expect(logger.info).toHaveBeenNthCalledWith(2, { name: 'CALLBACK' }, 'test callback: Request finished');
+      expect(fetch).toHaveBeenCalledWith('http://example.com/get/200/5/weird/content', { method: 'get', body: '{"about":"something"}', headers: { 'content-type': 'application/json' } });
+      expect(logger.info).toHaveBeenCalledTimes(2);
       expect(logger.error).not.toHaveBeenCalled();
       expect(logger.warn).not.toHaveBeenCalled();
     });