+++ packages/http/src/mocker/callback/callbacks.ts
@@ -1,7 +1,7 @@
 import { IHttpCallbackOperation, IHttpOperationRequest } from '@stoplight/types';
-import { resolveRuntimeExpressions } from '../../utils/runtimeExpression';
-import { IHttpRequest, IHttpResponse } from '../../types';
-import fetch from 'node-fetch';
+import fetch, { RequestInit } from 'node-fetch';
+import { Logger } from 'pino';
+import chalk from 'chalk';
 import * as Option from 'fp-ts/lib/Option';
 import * as Either from 'fp-ts/lib/Either';
 import { map, reduce } from 'fp-ts/lib/Array';
@@ -9,12 +9,15 @@ import * as TaskEither from 'fp-ts/lib/TaskEither';
 import * as ReaderTaskEither from 'fp-ts/lib/ReaderTaskEither';
 import { head } from 'fp-ts/lib/Array';
 import { pipe } from 'fp-ts/lib/pipeable';
+import { pick } from 'lodash';
+
+import { resolveRuntimeExpressions } from '../../utils/runtimeExpression';
+import { IHttpRequest, IHttpResponse } from '../../types';
 import { generate as generateHttpParam } from '../generator/HttpParamGenerator';
 import { validateOutput } from '../../validator';
 import { parseResponse } from '../../utils/parseResponse';
 import withLogger from '../../withLogger';
-import { violationLogger } from '../../utils/logger';
-import { Logger } from 'pino';
+import { logRequest, logResponse, violationLogger } from '../../utils/logger';
 
 export function runCallback({
   callback,
@@ -29,30 +32,43 @@ export function runCallback({
     const { url, requestData } = assembleRequest({ resource: callback, request, response });
     const logViolation = violationLogger(logger);
 
-    logger.info({ name: 'CALLBACK' }, `${callback.callbackName}: Making request to ${url}...`);
+    logCallbackRequest({ logger, callbackName: callback.callbackName, url, requestData });
 
     return pipe(
       TaskEither.tryCatch(() => fetch(url, requestData), Either.toError),
       TaskEither.chain(parseResponse),
+      TaskEither.map(callbackResponseLogger({ logger, callbackName: callback.callbackName })),
       TaskEither.mapLeft(error =>
-        logger.error({ name: 'CALLBACK' }, `${callback.callbackName}: Request failed: ${error.message}`)
+        logger.error(`${chalk.blueBright(callback.callbackName + ':')} Request failed: ${error.message}`)
       ),
-      TaskEither.chain(element => {
-        logger.info({ name: 'CALLBACK' }, `${callback.callbackName}: Request finished`);
-
-        return pipe(
-          validateOutput({ resource: callback, element }),
-          Either.mapLeft(violations => {
-            pipe(
-              violations,
-              map(logViolation)
-            );
-          }),
-          TaskEither.fromEither
-        );
-      })
+      TaskEither.chain(element => pipe(
+        validateOutput({ resource: callback, element }),
+        Either.mapLeft(violations => {
+          pipe(
+            violations,
+            map(logViolation)
+          );
+        }),
+        TaskEither.fromEither
+      ))
     );
   });
+}
+
+function logCallbackRequest({ logger, url, callbackName, requestData }: { logger: Logger, callbackName: string, url: string, requestData: Pick<RequestInit, 'headers' | 'method' | 'body'> }) {
+  const prefix = `${chalk.blueBright(callbackName + ':')} ${chalk.grey('> ')}`;
+  logger.info(`${prefix}Executing "${requestData.method}" callback to ${url}...`);
+  logRequest({ logger, prefix, ...pick(requestData, 'body', 'headers') });
+}
+
+function callbackResponseLogger({ logger, callbackName }: { logger: Logger, callbackName: string }) {
+  const prefix = `${chalk.blueBright(callbackName + ':')} ${chalk.grey('< ')}`;
+
+  return (response: IHttpResponse) => {
+    logger.info(`${prefix}Received callback response`);
+    logResponse({ logger, prefix, ...pick(response, 'body', 'headers', 'statusCode') });
+    return response;
+  };
 }
 
 function assembleRequest({